# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.DownloadListItem:
    properties:
      collection_id:
        type: integer
      download_url:
        type: string
      file_url:
        type: string
      format:
        type: string
      id:
        type: integer
      order:
        type: integer
      size_bytes:
        type: integer
      title:
        type: string
    type: object
  handlers.DownloadListResponse:
    properties:
      collection_id:
        type: integer
      items:
        items:
          $ref: '#/definitions/handlers.DownloadListItem'
        type: array
      total:
        type: integer
    type: object
  handlers.DownloadURLResponse:
    properties:
      collection_id:
        type: integer
      emoji_id:
        type: integer
      expires_at:
        type: integer
      key:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
  handlers.TelegramDownloadRequest:
    properties:
      link:
        type: string
      pack_name:
        type: string
    type: object
  handlers.TelegramDownloadResponse:
    properties:
      base_dir:
        type: string
      failed:
        type: integer
      files:
        items:
          $ref: '#/definitions/handlers.TelegramDownloadedFile'
        type: array
      pack_name:
        type: string
      saved:
        type: integer
      title:
        type: string
      total:
        type: integer
    type: object
  handlers.TelegramDownloadedFile:
    properties:
      emoji:
        type: string
      file_id:
        type: string
      file_path:
        type: string
      file_unique_id:
        type: string
      local_path:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Emoji collection site
  title: Emoji API
  version: 0.1.0
paths:
...
  /api/admin/telegram/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: telegram download
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TelegramDownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TelegramDownloadResponse'
      summary: Download Telegram sticker pack
      tags:
      - admin
...
  /api/collections/{id}/download-list:
    get:
      parameters:
      - description: collection id
        in: path
        name: id
        required: true
        type: integer
      - description: ttl (seconds)
        in: query
        name: ttl
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DownloadListResponse'
      summary: Get collection download list
      tags:
      - collections
  /api/collections/{id}/download-zip:
    get:
      parameters:
      - description: collection id
        in: path
        name: id
        required: true
        type: integer
      - description: ttl (seconds)
        in: query
        name: ttl
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DownloadURLResponse'
      summary: Get collection zip download URL
      tags:
      - collections
  /api/emojis/{id}/download:
    get:
      parameters:
      - description: emoji id
        in: path
        name: id
        required: true
        type: integer
      - description: ttl (seconds)
        in: query
        name: ttl
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DownloadURLResponse'
      summary: Get emoji download URL
      tags:
      - emojis
...
swagger: "2.0"